# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel E-commerce Store API
  description: |
    API RESTful completa para o Laravel E-commerce Store.

    Esta API oferece endpoints para:
    - Gestão de produtos, categorias e marcas
    - Carrinho de compras e pedidos
    - Sistema de avaliações e comentários
    - Gestão de clientes e endereços
    - Cupons e promoções
    - Relatórios e analytics
    - Administração completa

    ## Autenticação

    A API utiliza Laravel Sanctum para autenticação. Para acessar endpoints protegidos:

    1. Faça login para obter um token
    2. Inclua o token no header: `Authorization: Bearer {token}`
    3. Para operações administrativas, o usuário deve ter privilégios de admin

    ## Paginação

    Todos os endpoints que retornam listas suportam paginação:

    ```json
    {
      "success": true,
      "data": [...],
      "pagination": {
        "current_page": 1,
        "per_page": 15,
        "total": 100,
        "last_page": 7,
        "from": 1,
        "to": 15
      }
    }
    ```

    ## Filtros e Ordenação

    Use query parameters para filtrar e ordenar:

    - `?sort=name&order=asc` - Ordenar por nome (asc/desc)
    - `?category_id=1` - Filtrar por categoria
    - `?min_price=100&max_price=1000` - Filtrar por preço
    - `?search=smartphone` - Buscar produtos
    - `?in_stock=1` - Apenas produtos em estoque
    - `?featured=1` - Apenas produtos em destaque

    ## Cache

    A API utiliza cache inteligente com headers informativos:

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKey: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de autenticação obtido via `/api/v1/auth/login`.

        Inclua no header: `Authorization: Bearer {token}`
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Chave de API para acesso administrativo.

        Obtenha em: `/api/v1/admin/api-keys`
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Smartphone Samsung Galaxy"
        slug:
          type: string
          example: "smartphone-samsung-galaxy"
        sku:
          type: string
          example: "SAMSUNG-001"
        description:
          type: string
          example: "Smartphone de última geração com câmera de alta resolução"
        short_description:
          type: string
          example: "Smartphone Samsung Galaxy com 128GB"
        price:
          type: number
          format: float
          example: 1999.90
        compare_price:
          type: number
          format: float
          example: 2299.90
        cost:
          type: number
          format: float
          example: 1500.00
        stock_quantity:
          type: integer
          example: 50
        min_stock:
          type: integer
          example: 10
        max_stock:
          type: integer
          example: 1000
        is_active:
          type: boolean
          example: true
        is_featured:
          type: boolean
          example: true
        is_digital:
          type: boolean
          example: false
        weight:
          type: number
          format: float
          example: 0.5
        length:
          type: number
          format: float
          example: 15.0
        width:
          type: number
          format: float
          example: 7.0
        height:
          type: number
          format: float
          example: 1.0
        meta_title:
          type: string
          example: "Smartphone Samsung Galaxy - Loja Online"
        meta_description:
          type: string
          example: "Compre o Smartphone Samsung Galaxy com 128GB de armazenamento"
        tags:
          type: array
          items:
            type: string
          example: ["smartphone", "samsung", "galaxy"]
        category:
          $ref: '#/components/schemas/Category'
        brand:
          $ref: '#/components/schemas/Brand'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        average_rating:
          type: number
          format: float
          example: 4.5
        total_reviews:
          type: integer
          example: 25
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erro de validação"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            name: ["O campo nome é obrigatório"]
            email: ["O campo email deve ser um email válido"]
        code:
          type: integer
          example: 422
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
          example: "Operação realizada com sucesso"
        pagination:
          $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
...
paths:
  /api/v1/products:
    get:
      tags:
        - Products
      summary: Listar produtos
      description: |
        Retorna uma lista paginada de produtos com opções de filtro e ordenação.

        ### Filtros Disponíveis:
        - `category_id` - Filtrar por categoria
        - `brand_id` - Filtrar por marca
        - `min_price`, `max_price` - Filtrar por preço
        - `min_rating` - Filtrar por avaliação mínima
        - `in_stock` - Apenas produtos em estoque (1) ou sem estoque (0)
        - `featured` - Apenas produtos em destaque (1)
        - `on_sale` - Apenas produtos em promoção (1)
        - `search` - Buscar por nome, descrição ou SKU

        ### Ordenação:
        - `sort` - Campo para ordenar (name, price, rating, created_at)
        - `order` - Direção (asc, desc)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Página atual
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 15
          description: Itens por página
        - name: category_id
          in: query
          schema:
            type: integer
          description: ID da categoria
        - name: brand_id
          in: query
          schema:
            type: integer
          description: ID da marca
        - name: min_price
          in: query
          schema:
            type: number
            format: float
          description: Preço mínimo
        - name: max_price
          in: query
          schema:
            type: number
            format: float
          description: Preço máximo
        - name: min_rating
          in: query
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 5
          description: Avaliação mínima
        - name: in_stock
          in: query
          schema:
            type: integer
            enum: [0, 1]
          description: Status de estoque
        - name: featured
          in: query
          schema:
            type: integer
            enum: [0, 1]
          description: Produtos em destaque
        - name: on_sale
          in: query
          schema:
            type: integer
            enum: [0, 1]
          description: Produtos em promoção
        - name: search
          in: query
          schema:
            type: string
            minLength: 2
          description: Termo de busca
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, price, rating, created_at]
            default: name
          description: Campo para ordenar
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Direção da ordenação
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Muitas requisições
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Products
      summary: Criar produto (Admin)
      description: Cria um novo produto. Requer autenticação de administrador.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - stock_quantity
                - category_id
                - brand_id
              properties:
                category_id:
                  type: integer
                  example: 1
                brand_id:
                  type: integer
                  example: 1
                name:
                  type: string
                  minLength: 3
                  maxLength: 255
                  example: "Smartphone Samsung Galaxy"
                slug:
                  type: string
                  example: "smartphone-samsung-galaxy"
                sku:
                  type: string
                  example: "SAMSUNG-001"
                description:
                  type: string
                  example: "Smartphone de última geração"
                short_description:
                  type: string
                  example: "128GB de armazenamento"
                price:
                  type: number
                  format: float
                  minimum: 0
                  example: 1999.90
                compare_price:
                  type: number
                  format: float
                  example: 2299.90
                cost:
                  type: number
                  format: float
                  example: 1500.00
                stock_quantity:
                  type: integer
                  minimum: 0
                  example: 50
                min_stock:
                  type: integer
                  example: 10
                max_stock:
                  type: integer
                  example: 1000
                is_active:
                  type: boolean
                  example: true
                is_featured:
                  type: boolean
                  example: false
                is_digital:
                  type: boolean
                  example: false
                weight:
                  type: number
                  format: float
                  example: 0.5
                length:
                  type: number
                  format: float
                  example: 15.0
                width:
                  type: number
                  format: float
                  example: 7.0
                height:
                  type: number
                  format: float
                  example: 1.0
                meta_title:
                  type: string
                  example: "Smartphone Samsung Galaxy - Loja Online"
                meta_description:
                  type: string
                  example: "Compre o melhor smartphone Samsung"
                tags:
                  type: array
                  items:
                    type: string
                  example: ["smartphone", "samsung", "galaxy"]
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/admin/products:
    post:
      tags:
        - Admin
      summary: Criar produto (Admin)
      description: Cria um novo produto. Requer privilégios de administrador.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - stock_quantity
                - category_id
                - brand_id
              properties:
                category_id:
                  type: integer
                  example: 1
                brand_id:
                  type: integer
                  example: 1
                name:
                  type: string
                  minLength: 3
                  maxLength: 255
                  example: "Smartphone Samsung Galaxy"
                slug:
                  type: string
                  example: "smartphone-samsung-galaxy"
                sku:
                  type: string
                  example: "SAMSUNG-001"
                description:
                  type: string
                  example: "Smartphone de última geração"
                short_description:
                  type: string
                  example: "128GB de armazenamento"
                price:
                  type: number
                  format: float
                  minimum: 0
                  example: 1999.90
                compare_price:
                  type: number
                  format: float
                  example: 2299.90
                cost:
                  type: number
                  format: float
                  example: 1500.00
                stock_quantity:
                  type: integer
                  minimum: 0
                  example: 50
                min_stock:
                  type: integer
                  example: 10
                max_stock:
                  type: integer
                  example: 1000
                is_active:
                  type: boolean
                  example: true
                is_featured:
                  type: boolean
                  example: false
                is_digital:
                  type: boolean
                  example: false
                weight:
                  type: number
                  format: float
                  example: 0.5
                length:
                  type: number
                  format: float
                  example: 15.0
                width:
                  type: number
                  format: float
                  example: 7.0
                height:
                  type: number
                  format: float
                  example: 1.0
                meta_title:
                  type: string
                  example: "Smartphone Samsung Galaxy - Loja Online"
                meta_description:
                  type: string
                  example: "Compre o melhor smartphone Samsung"
                tags:
                  type: array
                  items:
                    type: string
                  example: ["smartphone", "samsung", "galaxy"]
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...